fix: prove the database topology and fencing per member
Address the second review wave on PR #345.

* Stop truncating collector evidence with [:4]; the 4096-byte output guard
  now fails closed instead of silently hiding SBD devices or instances that
  would change the classification.
* Emit hana.installed and hana.sr_online so a Pacemaker majority maker (a
  member that runs no HANA) is distinguishable from a HANA node whose system
  replication is down.
* _db_details partitions members into HANA nodes and witnesses, requires
  replication online on every HANA node and at least two of them, and renders
  only the HANA nodes, so a supported scale-out topology is generatable.
* _cluster_type rejects a member that reports no fencing agents, and rejects
  members that disagree, instead of unioning their evidence.
* _unique_instances compares virtual_host as part of the canonical peer
  projection.
* Split the 1288-line test module into shared fixtures, classification tests,
  and publication tests so every module stays under the 1000-line lint limit.

Copilot-Session: 0b9ff954-e372-4deb-86b2-a6dcc035d438
